### 2026-01-20 #9867 @crvernon
@amsks - Thank you for your submission to JOSS.  Please address the following issues:

AND any DOI issues listed here:  https://github.com/openjournals/joss-reviews/issues/9867#issuecomment-3772824420

Let me know when these have been addressed.
