### 2026-02-19 #9867 @amsks
Hi @crvernon

Thank you for your comment. We made the following changes in the commit [ef695d21b97a8810e379054f20317ef8409f5aa9](https://github.com/automl/Mighty/commit/ef695d21b97a8810e379054f20317ef8409f5aa9):
- Added a section on software design in paper.md
- Added impact and AI disclosure statements to paper.md
- Added the DOI's mentioned in [#9867](https://github.com/openjournals/joss-reviews/issues/9867#issuecomment-3772824420)

Hope this address the checks.
