### 2026-03-22 #9867 @abhishektiwari
@amsks  It seems your change is still not applied to `joss_paper` branch. Maybe you need to [merge your change](https://github.com/automl/Mighty/compare/joss_paper...Aditya-Patch-1) to `joss_paper` branch?

https://github.com/automl/Mighty/tree/joss_paper/paper
