### 2026-03-22 #9867 @amsks
@abhishektiwari I have merged the changes to the joss_paper branch (https://github.com/automl/Mighty/commit/b768d932afde6b10ebc53944c91cf1f2d48bb152) -- I hope the check passes now.
