### 2026-03-22 #9867 @abhishektiwari
Thanks @amsks

I see [project is using](https://github.com/automl/Mighty/?tab=License-1-ov-file) some customisation of [The 3-Clause BSD License](https://opensource.org/license/bsd-3-clause) by including,

```
All rights reserved.
```

Is this intentional? Some more context will be helpful.
