### 2026-03-23 #9867 @amsks
@abhishektiwari Thanks for pointing this out. I have standardised the BDS-3 license by removing that line.
