### 2026-03-23 #9867 @abhishektiwari
Thanks @amsks. Let make sure above License change  applied to `main` branch prior to the review process.

Can you suggest a few potential reviewers? You can also share GitHub handles of potential reviewers. Please be sure not to @-tag your suggestions, to avoid unwanted GitHub notifications to potential reviewers. Please also read the our [conflict of interest](https://github.com/openjournals/joss/blob/main/COI.md) (COI) when recommending reviewers. You can find reviewers in our [reviewers database](https://reviewers.joss.theoj.org/) or authors/reviewers of most similar historical JOSS papers on this issue.
