### 2026-04-06 #9867 @abhishektiwari
👋 @janikmu @tomzhengy @ab93, would any of you be willing to review this submission for Journal of Open Source Software (JOSS)?  The submission under consideration is "Mighty: A Comprehensive Tool for studying Generalization, Meta-RL and AutoRL".

Thanks
Abhi
