### 2026-04-17 #9867 @abhishektiwari
@kvr06-ai @amsks @MarcusBinder @ymahlau, would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "Mighty: A Comprehensive Tool for studying Generalization, Meta-RL and AutoRL".

Thanks
Abhi
