### 2026-04-18 #9867 @kvr06-ai
@abhishektiwari, quick process question before I start the review. Per the [JOSS Editorial Guide](https://joss.readthedocs.io/en/latest/editing.html) (steps 9–10), the standard flow is to run `@editorialbot start review` to open a dedicated review issue with per-reviewer checklists, then close this pre-review. Would you like to follow that path, or would you prefer review comments go directly in this pre-review thread? Happy to do it either way - just want to make sure I file in the right place.
