### 2026-04-19 #10439 @editorialbot
[REVIEW]: Mighty: A Comprehensive Tool for studying Generalization, Meta-RL and AutoRL
**Submitting author:** @amsks (<a href="http://orcid.org/0000-0003-0092-3780">Aditya Mohan</a>)
**Repository:** https://github.com/automl/Mighty/
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v1.0.0
