### 2026-04-20 #10439 @kvr06-ai
## First-pass review

Setup: installed `mighty-rl@v1.0.0` into a clean Python 3.11 `uv` venv, ran the pytest suite, smoke-tested `run_mighty.py` (PPO and DQN on CartPole, plus PPO with the RND meta-component), ran `mkdocs build --strict`, and read the PDF proof.

Recommending **Minor Revisions**. The four items below are small and easy to address.

### Checklist items not yet tickable

**1. Installation** (*"Does installation proceed as outlined in the documentation?"*)

`uv pip install mighty-rl` (and `make install`) resolves `gymnasium` to `1.2.3`. That breaks `mighty/mighty_agents/base_agent.py:33`, which imports `NormalizeObservation` and `NormalizeReward` from `gymnasium.wrappers.normalize` (removed in gymnasium 1.0). Result: 15 test modules fail at collection and `import mighty` raises `ModuleNotFoundError`. Pinning `gymnasium<1.0` or updating the import restores a working install. Related: #103, #108.

**2. Community guidelines** (*"Are there clear guidelines for third parties wishing to 1. Contribute… 2. Report issues… 3. Seek support"*)

The repo has no `CONTRIBUTING.md`. The PR template at `.github/Pull_Request_Template/pull_request_template.md:3` already references the missing file ("please ensure you have taken a look at the contribution guidelines in the CONTRIBUTING.md"). Issue reporting is covered by the issue template, but contribution and support pathways are not.

**3. Good practices** (*"…clear contribution/support pathways"*)

Same gap as item 2. This item becomes tickable once `CONTRIBUTING.md` (and ideally `CODE_OF_CONDUCT.md`) is added.

**4. Reproducibility** (*"original results are entirely reproducible by reviewers"*)

The CartPole and MountainCar rows in the validation table reproduce from committed configs. `mighty/configs/environment/gymnasium/` contains cartpole, mountaincar, mountaincarcontinuous, pendulum, and atari_pong only. The SAC rows on Walker2D and HalfCheetah are therefore not directly reproducible. Committing the configs used for those runs would resolve this.

### Observations that do not cleanly map to a checklist item

These came up during review but do not map to a specific box.

- **`docs/index.md:73-75`** calls Mighty "early development" and lists Meta-RL and AutoRL as future work, which contradicts the paper. Does not map to *Statement of need (docs)*: the section itself is clear and complete (the issue is editorial drift in one paragraph, not a missing element).
- **`10.52202/068431-1197`** (editorialbot's candidate DOI for Lu et al., `lu-neurips22a`) is worth verifying and adding. Does not map to *References*: the bibliography is complete with proper citation syntax, and an optional DOI is not itself a checklist violation.
- **"Cite Us" BibTeX** in `README.md` and `docs/index.md:83-88` disagree on authors (docs adds F. Hutter). Does not map to *References* or *Contribution and authorship*: this concerns how users cite the software, not the paper's bibliography or its byline.

### Positives

- All required paper sections present (Summary, Statement of need, State of the field, Software Design, Research Impact, AI disclosure); word count 1758.
- State of the field positions Mighty against TorchRL, Tianshou, SB3, CleanRL, RLlib, and STOIX with an explicit build-vs-contribute argument.
- BSD-3 (OSI) license; sustained development since May 2021; v1.0.0 released August 2025; one external PR contributor (@jmtoepperwien).
- `mkdocs build --strict` runs clean; API docs are auto-generated via mkdocstrings.
- With `gymnasium` pinned, the framework works end-to-end: PPO, DQN, and PPO + RND each train cleanly on CartPole within seconds.

**Issue tracking**: filed automl/Mighty#116 for items 2–3 (CONTRIBUTING.md / CODE_OF_CONDUCT). Item 1 (gymnasium) is already tracked upstream in #103 and #108. Item 4 (reproducibility) is small enough to address directly from this thread.
