### 2026-04-20 automl/Mighty#116 @kvr06-ai
Add CONTRIBUTING.md and CODE_OF_CONDUCT.md (JOSS review blocker)
Flagged during the JOSS review at openjournals/joss-reviews#10439.

JOSS's community-guidelines checklist requires clear guidelines for (1) contributing, (2) reporting issues, and (3) seeking support. The repo currently has no `CONTRIBUTING.md`, and the PR template at `.github/Pull_Request_Template/pull_request_template.md:3` already references the missing file ("please ensure you have taken a look at the contribution guidelines in the CONTRIBUTING.md").

Adding `CONTRIBUTING.md` (and ideally `CODE_OF_CONDUCT.md`) unblocks two JOSS checklist items: *Community guidelines*, and *Good practices* (which explicitly calls for "clear contribution/support pathways").

Minimum content suggestion for `CONTRIBUTING.md`:
- How to set up a dev environment (linking `README.md` and the `Makefile` is fine).
- How to run tests and pre-commit locally.
- PR expectations (aligned with the existing PR template).
- Where to seek support (issues, discussions, contact email).
