### 2026-05-28 2759cfb
fix(exploration): typo self.ago -> self.algo in sample_func_logits
In MightyExplorationPolicy.sample_func_logits, the sac flag passed to
sample_nondeterministic_logprobs used self.ago instead of self.algo.
Because self.ago does not exist, this always resolved to False, meaning
the tanh-squash correction was never applied for SAC in this code path.
