+++ mighty/mighty_exploration/mighty_exploration_policy.py
@@ -111,7 +111,7 @@ class MightyExplorationPolicy:
         elif isinstance(out, tuple) and len(out) == 4:
             action = out[0]  # [batch, action_dim]
             log_prob = sample_nondeterministic_logprobs(
-                z=out[1], mean=out[2], log_std=out[3], sac=self.ago == "sac"
+                z=out[1], mean=out[2], log_std=out[3], sac=self.algo == "sac"
             )
             return action.detach().cpu().numpy(), log_prob
 
