### 2026-07-12 e68faf2
chore: bump version to 1.0.1 for release
