### 2026-07-13 6d1c69d
docs: use André Biedenkapp (accent) to match ORCID record
