### 2026-07-13 b73126c
docs: reword README intro to drop early-stage framing (JOSS review)
