### 2026-07-14 2b4b107
docs: address reviewer reference corrections (JOSS review)
Apply the copy-edits requested in openjournals/joss-reviews#10439:

paper.md
- "stable baselines3" -> "Stable-baselines3" to match the reference
- "OpenRL benchmark" -> "Open RL Benchmark" (OpenRL is a different
  library); also corrected in the results table and discussion
- "Pufferlib" -> "PufferLib"

paper.bib
- brace-protect proper nouns so JOSS sentence-casing does not mangle
  them (Atari, TorchRL/PyTorch, Q-learning, CleanRL, Open RL Benchmark,
  JAX, AutoRL, Go, PufferLib, Gran Turismo, NovelD, and others)
- fix "Science in Rob
[…]
