+++ paper/paper.md
@@ -46,18 +46,18 @@ Robust generalization, rapid adaptation, and automated tuning are critical for d
 
 Reinforcement learning (RL) has emerged as a powerful decision-making paradigm in complex and dynamic environments. Despite impressive successes in domains such as games [@silver-nature16a; @badia-icml20a; @vasco-rlc24] and robotics [@lee-sciro20a], RL algorithms frequently overfit their training conditions and struggle to generalize to new tasks [@benjamins-tmlr23a; @kirk-jair23a; @mohan-jair24a]. Addressing this challenge requires methods that not only learn efficiently on a single task but also adapt rapidly to novel settings and automatically tune their learning process.
 
-Recent research has advanced in three complementary directions: (i) Generalization in RL [@benjamins-tmlr23a; @cho-neurips24a; @mohan-jair24a], (ii) Meta-RL methods [@kaushik-iros20a; @beck-arxiv23a], and (iii) Automated RL (AutoRL) [@parkerholder-jair22a; @mohan-automlconf23a; @eimer-icml23a]. Although each has led to promising algorithms, researchers frequently resort to fragmented codebases and ad hoc scripting across environment design, RL training, and meta-optimization. This fragmentation increases engineering effort, impedes rapid iteration, and undermines reproducibility [@paradis-rlc].
+Recent research has advanced in three complementary directions: (i) Generalization in RL [@benjamins-tmlr23a; @cho-neurips24a; @mohan-jair24a], (ii) Meta-RL methods [@kaushik-iros20a; @beck-fntml25a], and (iii) Automated RL (AutoRL) [@parkerholder-jair22a; @mohan-automlconf23a; @eimer-icml23a]. Although each has led to promising algorithms, researchers frequently resort to fragmented codebases and ad hoc scripting across environment design, RL training, and meta-optimization. This fragmentation increases engineering effort, impedes rapid iteration, and undermines reproducibility [@paradis-rlc].
 
 We introduce *Mighty*: a modular library designed to enable research at the intersection of generalization, Meta-RL, and AutoRL. Mighty enforces a clean and principled separation between inner- and outer-loop processes, making it easy to combine, for example, curricula, context adaptation, and automated tuning within a unified framework. Users can prototype new methods, compose existing ones, and run controlled comparisons - all without ad hoc orchestration code.
 
 ![Overview of Mighty's concept and modules.](Figures/mighty_concept.pdf)
 
-Mighty is designed around three design principles: *flexibility, smooth integration with existing libraries, and environment parallelization*. First, flexibility is central. Mighty exposes transitions, predictions, networks, and environments to meta-methods, enabling a broad range of research patterns including black-box outer loops, algorithm-informed inner loops, and environment-level interventions. Second, Mighty integrates smoothly with Gymnasium [@towers-arxiv24a], Pufferlib [@suarez-rlc25], CARL [@benjamins-tmlr23a], and can interface with tools such as evosax [@evosax2022github] in under $100$ lines of code. This minimizes the glue code while preserving flexibility. Finally, Mighty uses standard Python and PyTorch for optimized networks with vectorized CPU environments for fast environment interaction. This design offers high training speeds, even for purely CPU-based environments, without sacrificing algorithmic modularity or code clarity.
[…]
