+++ paper/paper.bib
@@ -1,32 +1,41 @@
 @inproceedings{badia-icml20a,
-  title={Agent57: Outperforming the atari human benchmark},
+  title={{Agent57}: Outperforming the {Atari} human benchmark},
   author={Badia, A. and Piot, B. and Kapturowski, S. and Sprechmann, P. and Vitvitskyi, A. and Guo, Z. and Blundell, C.},
   booktitle={Proceedings of the 37th International Conference on Machine Learning (ICML'20)},
-  volume={98},
-  editor={Daume III, H. and Singh, A.},
-  publisher={Proceedings of Machine Learning Research},
-  year={2020}
+  volume={119},
+  pages={507--517},
+  editor={Daum{\'e} III, H. and Singh, A.},
+  publisher={PMLR},
+  series={Proceedings of Machine Learning Research},
+  year={2020},
+  url={https://proceedings.mlr.press/v119/badia20a.html}
 }
 
-@article{beck-arxiv23a,
-  title={A survey of meta-reinforcement learning},
+@article{beck-fntml25a,
+  title={A tutorial on meta-reinforcement learning},
   author={Beck, J. and Vuorio, R. and Liu, E. and Xiong, Z. and Zintgraf, L. and Finn, C. and Whiteson, S.},
-  journal={CoRR},
-  volume={abs/2301.08028},
-  year={2023}
+  journal={Foundations and Trends in Machine Learning},
+  volume={18},
+  number={2--3},
+  pages={224--384},
+  publisher={Now Publishers},
+  year={2025},
+  doi={10.1561/2200000080}
 }
 
 @article{benjamins-tmlr23a,
   title={Contextualize me – the case for context in reinforcement learning},
   author={Benjamins, C. and Eimer, T. and Schubert, F. and Mohan, A. and D{\"o}hler, S. and Biedenkapp, A. and Rosenhan, B. and Hutter, F. and Lindauer, M.},
   journal={Transactions on Machine Learning Research},
-  year={2023}
+  year={2023},
+  url={https://openreview.net/forum?id=Y42xVBQusn}
 }
 
 @misc{bou-arxiv23a,
-  title={Torchrl: A data-driven decision-making library for pytorch},
+  title={{TorchRL}: A data-driven decision-making library for {PyTorch}},
   author={Bou, A. and Bettini, M. and Dittert, S. and Kumar, V. and Sodhani, S. and Yang, X. and De Fabritiis, G. and Moens, V.},
-  year={2023}
+  year={2023},
+  doi={10.48550/arXiv.2306.00577}
 }
 
 @inproceedings{burda-iclr19a,
@@ -35,7 +44,8 @@
   booktitle={The Seventh International Conference on Learning Representations (ICLR'19)},
   publisher={ICLR},
   year={2019},
-  note={Published online: iclr.cc}
+  note={Published online: iclr.cc},
+  url={https://openreview.net/forum?id=rJNwDjAqYX}
 }
 
 @inproceedings{cho-neurips24a,
@@ -53,15 +63,18 @@
   booktitle={The Ninth International Conference on Learning Representations (ICLR'21)},
   publisher={ICLR},
   year={2021},
-  note={Published online: iclr.cc}
+  note={Published online: iclr.cc},
+  url={https://openreview.net/forum?id=ONBPHFZ7zG4}
 }
 
 @article{paradis-rlc,
   title={Resource usage evaluation of discrete model-free deep reinforcement learning algorithms},
   author={Dizon-Paradis, O. and Wormald, S. and Capecci, D. and Bhandarkar, A. and Woodard, D.},
   journal={Reinforcement Learning Journal},
+  volume={5},
+  pages={2162--2177},
   year={2024},
[…]
