### 2026-07-23 6047de0
Update acknowledgements and AI usage disclosure
Added acknowledgements for funding contributions and AI usage disclosure.
