### 2026-07-26 7111024
docs: address final JOSS review edits (title case, epsilon-z-greedy note, bib formatting)
