+++ paper/paper.bib
@@ -351,8 +351,8 @@
 
 @inproceedings{hasselt-aaai16a,
   title={Deep reinforcement learning with double {Q}-learning},
-  author={van Hasselt, H. and Guez, A. and Silver, D.},
-  booktitle={Proceedings of the Thirtieth AAAI Conference on Artificial Intelligence (AAAI'16)},
+  author={{van Hasselt}, H. and Guez, A. and Silver, D.},
+  booktitle={{Proceedings of the Thirtieth AAAI Conference on Artificial Intelligence (AAAI'16)}},
   pages={2094--2100},
   editor={Schuurmans, D. and Wellman, M.},
   publisher={AAAI Press},
