### 2026-08-01 2c63123
docs: add JOSS DOI badge, published citation, and CITATION.cff
- add JOSS status badge (10.21105/joss.10439) to README badge block
- replace placeholder @misc citation with the published @article in README and docs/index.md
- add CITATION.cff so GitHub's "Cite this repository" menu resolves to the paper
